Implement #356
