+++ .flake8
@@ -11,3 +11,7 @@ extend-ignore =
             # Missing type annotation for self in method
             N815
             # camelCase
+per-file-ignores =
+    # long urls in test fixtures cannot be avoided, and we want them to be on one line for
+    # legibility reasons.
+    database/test/conftest.py:E501
