+++ database/test/test_ryhti_client.py
@@ -1,401 +1,16 @@
 import json
 import re
-from typing import Callable, List, Mapping, Optional
+from typing import Callable
 
 import codes
 import models
 import pytest
-from base import PROJECT_SRID
 from requests_mock.request import _RequestObjectProxy
 from ryhti_client.ryhti_client import RyhtiClient
 from simplejson import JSONEncoder
 from sqlalchemy.orm import Session
 
-
-@pytest.fixture(scope="function")
-def desired_plan_dict(
-    plan_instance: models.Plan,
-    land_use_area_instance: models.LandUseArea,
-    land_use_point_instance: models.LandUsePoint,
-    plan_regulation_group_instance: models.PlanRegulationGroup,
-    point_plan_regulation_group_instance: models.PlanRegulationGroup,
-    general_regulation_group_instance: models.PlanRegulationGroup,
-    empty_value_plan_regulation_instance: models.PlanRegulation,
-    text_plan_regulation_instance: models.PlanRegulation,
-    point_text_plan_regulation_instance: models.PlanRegulation,
-    numeric_plan_regulation_instance: models.PlanRegulation,
-    verbal_plan_regulation_instance: models.PlanRegulation,
-    general_plan_regulation_instance: models.PlanRegulation,
-    plan_proposition_instance: models.PlanProposition,
-    pending_date_instance: models.LifeCycleDate,
-) -> dict:
-    """
-    Plan dict based on https://github.com/sykefi/Ryhti-rajapintakuvaukset/blob/main/OpenApi/Kaavoitus/Avoin/ryhti-plan-public-validate-api.json
-
-    Let's 1) write explicitly the complex fields, and 2) just check that the simple fields have
-    the same values as the original plan fixture in the database.
-    """
-
-    return {
-        "planKey": plan_instance.id,
-        "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-        "scale": plan_instance.scale,
-        "geographicalArea": {
-            "srid": str(PROJECT_SRID),
-            "geometry": {
-                "type": "Polygon",
-                "coordinates": [
-                    [
-                        [381849.834412134019658, 6677967.973336197435856],
-                        [381849.834412134019658, 6680613.389312859624624],
-                        [386378.427863708813675, 6680613.389312859624624],
-                        [386378.427863708813675, 6677967.973336197435856],
-                        [381849.834412134019658, 6677967.973336197435856],
-                    ]
-                ],
-            },
-        },
-        # TODO: plan documents to be added.
-        "periodOfValidity": None,
-        "approvalDate": None,
-        "generalRegulationGroups": [
-            {
-                "generalRegulationGroupKey": general_regulation_group_instance.id,
-                "titleOfPlanRegulation": general_regulation_group_instance.name,
-                "planRegulations": [
-                    {
-                        "planRegulationKey": general_plan_regulation_instance.id,
-                        "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                        "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayslaji/code/asumisenAlue",
-                        "value": {
-                            "dataType": "LocalizedText",
-                            "text": general_plan_regulation_instance.text_value,
-                        },
-                        "subjectIdentifiers": [
-                            general_plan_regulation_instance.name[
-                                "fin"
-                            ]  # TODO: onko asiasana aina yksikielinen??
-                        ],
-                        "additionalInformations": [
-                            {
-                                "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayksen_Lisatiedonlaji/code/paakayttotarkoitus"
-                            }
-                        ],
-                        "planThemes": [
-                            "http://uri.suomi.fi/codelist/rytj/kaavoitusteema/code/01",
-                        ],
-                        # oh great, integer has to be string here for reasons unknown.
-                        "regulationNumber": str(
-                            general_plan_regulation_instance.ordering
-                        ),
-                        # TODO: plan regulation documents to be added.
-                        "periodOfValidity": None,
-                    },
-                ],
-                "planRecommendations": [],
-            }
-        ],
-        "planDescription": plan_instance.description[
-            "fin"
-        ],  # TODO: should this be a single language string? why?
-        "planObjects": [
-            {
-                "planObjectKey": land_use_area_instance.id,
-                "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                "undergroundStatus": "http://uri.suomi.fi/codelist/rytj/RY_MaanalaisuudenLaji/code/01",
-                "geometry": {
-                    "srid": str(PROJECT_SRID),
-                    "geometry": {
-                        "type": "Polygon",
-                        "coordinates": [
-                            [
-                                [381849.834412134019658, 6677967.973336197435856],
-                                [381849.834412134019658, 6680613.389312859624624],
-                                [386378.427863708813675, 6680613.389312859624624],
-                                [386378.427863708813675, 6677967.973336197435856],
-                                [381849.834412134019658, 6677967.973336197435856],
-                            ]
-                        ],
-                    },
-                },
-                "name": land_use_area_instance.name,
-                "description": land_use_area_instance.description,
-                "objectNumber": land_use_area_instance.ordering,
-                "verticalLimit": {
-                    "dataType": "decimalRange",
-                    "minimumValue": land_use_area_instance.height_range.lower,
-                    "maximumValue": land_use_area_instance.height_range.upper,
-                    "unitOfMeasure": land_use_area_instance.height_unit,
-                },
-                "periodOfValidity": None,
-            },
-            {
-                "planObjectKey": land_use_point_instance.id,
-                "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                "undergroundStatus": "http://uri.suomi.fi/codelist/rytj/RY_MaanalaisuudenLaji/code/01",
-                "geometry": {
-                    "srid": str(PROJECT_SRID),
-                    "geometry": {
-                        "type": "Point",
-                        "coordinates": [382000, 6678000],
-                    },
-                },
-                "name": land_use_point_instance.name,
-                "description": land_use_point_instance.description,
-                "objectNumber": land_use_point_instance.ordering,
-                "periodOfValidity": None,
-            },
-        ],
-        "planRegulationGroups": [
-            {
-                "planRegulationGroupKey": point_plan_regulation_group_instance.id,
-                "titleOfPlanRegulation": point_plan_regulation_group_instance.name,
-                "planRegulations": [
-                    {
-                        "planRegulationKey": point_text_plan_regulation_instance.id,
-                        "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                        "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayslaji/code/asumisenAlue",
-                        "value": {
-                            "dataType": "LocalizedText",
-                            "text": point_text_plan_regulation_instance.text_value,
-                        },
-                        "subjectIdentifiers": [
-                            point_text_plan_regulation_instance.name[
-                                "fin"
-                            ]  # TODO: onko asiasana aina yksikielinen??
-                        ],
-                        "additionalInformations": [
-                            {
-                                "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayksen_Lisatiedonlaji/code/paakayttotarkoitus"
-                            }
-                        ],
-                        "planThemes": [
-                            "http://uri.suomi.fi/codelist/rytj/kaavoitusteema/code/01",
-                        ],
-                        # oh great, integer has to be string here for reasons unknown.
-                        "regulationNumber": str(
-                            point_text_plan_regulation_instance.ordering
-                        ),
-                        # TODO: plan regulation documents to be added.
-                        "periodOfValidity": None,
-                    }
-                ],
-                "planRecommendations": [],
-                "letterIdentifier": point_plan_regulation_group_instance.short_name,
-                "colorNumber": "#FFFFFF",
-            },
-            {
-                "planRegulationGroupKey": plan_regulation_group_instance.id,
-                "titleOfPlanRegulation": plan_regulation_group_instance.name,
-                "planRegulations": [
-                    {
-                        "planRegulationKey": empty_value_plan_regulation_instance.id,
-                        "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                        "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayslaji/code/asumisenAlue",
-                        "subjectIdentifiers": [
-                            empty_value_plan_regulation_instance.name[
-                                "fin"
-                            ]  # TODO: onko asiasana aina yksikielinen??
-                        ],
-                        "additionalInformations": [
-                            {
-                                "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayksen_Lisatiedonlaji/code/paakayttotarkoitus"
-                            }
-                        ],
-                        "planThemes": [
-                            "http://uri.suomi.fi/codelist/rytj/kaavoitusteema/code/01",
-                        ],
-                        # oh great, integer has to be string here for reasons unknown.
-                        "regulationNumber": str(
-                            empty_value_plan_regulation_instance.ordering
-                        ),
-                        # TODO: plan regulation documents to be added.
-                        "periodOfValidity": None,
-                    },
-                    {
-                        "planRegulationKey": numeric_plan_regulation_instance.id,
-                        "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                        "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayslaji/code/asumisenAlue",
-                        "value": {
-                            "dataType": "decimal",
-                            "number": numeric_plan_regulation_instance.numeric_value,
-                            "unitOfMeasure": numeric_plan_regulation_instance.unit,
-                        },
-                        "subjectIdentifiers": [
-                            numeric_plan_regulation_instance.name[
-                                "fin"
-                            ]  # TODO: onko asiasana aina yksikielinen??
-                        ],
-                        "additionalInformations": [
-                            {
-                                "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayksen_Lisatiedonlaji/code/paakayttotarkoitus"
-                            }
-                        ],
-                        "planThemes": [
-                            "http://uri.suomi.fi/codelist/rytj/kaavoitusteema/code/01",
-                        ],
-                        # oh great, integer has to be string here for reasons unknown.
-                        "regulationNumber": str(
-                            numeric_plan_regulation_instance.ordering
-                        ),
-                        # TODO: plan regulation documents to be added.
-                        "periodOfValidity": None,
-                    },
-                    {
-                        "planRegulationKey": text_plan_regulation_instance.id,
-                        "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                        "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayslaji/code/asumisenAlue",
-                        "value": {
-                            "dataType": "LocalizedText",
-                            "text": text_plan_regulation_instance.text_value,
-                        },
-                        "subjectIdentifiers": [
-                            text_plan_regulation_instance.name[
-                                "fin"
-                            ]  # TODO: onko asiasana aina yksikielinen??
-                        ],
-                        "additionalInformations": [
-                            {
-                                "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayksen_Lisatiedonlaji/code/paakayttotarkoitus"
-                            }
-                        ],
-                        "planThemes": [
-                            "http://uri.suomi.fi/codelist/rytj/kaavoitusteema/code/01",
-                        ],
-                        # oh great, integer has to be string here for reasons unknown.
-                        "regulationNumber": str(text_plan_regulation_instance.ordering),
-                        # TODO: plan regulation documents to be added.
-                        "periodOfValidity": None,
-                    },
-                    {
-                        "planRegulationKey": verbal_plan_regulation_instance.id,
-                        "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                        "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayslaji/code/sanallinenMaarays",
-                        "value": {
-                            "dataType": "LocalizedText",
-                            "text": verbal_plan_regulation_instance.text_value,
-                        },
-                        "subjectIdentifiers": [
-                            verbal_plan_regulation_instance.name[
-                                "fin"
-                            ]  # TODO: onko asiasana aina yksikielinen??
-                        ],
-                        "verbalRegulations": [
-                            "http://uri.suomi.fi/codelist/rytj/RY_Sanallisen_Kaavamaarayksen_Laji/code/perustaminen"
-                        ],
-                        "additionalInformations": [
-                            {
-                                "type": "http://uri.suomi.fi/codelist/rytj/RY_Kaavamaarayksen_Lisatiedonlaji/code/paakayttotarkoitus"
-                            }
-                        ],
-                        "planThemes": [
-                            "http://uri.suomi.fi/codelist/rytj/kaavoitusteema/code/01",
-                        ],
-                        # oh great, integer has to be string here for reasons unknown.
-                        "regulationNumber": str(
-                            verbal_plan_regulation_instance.ordering
-                        ),
-                        # TODO: plan regulation documents to be added.
-                        "periodOfValidity": None,
-                    },
-                ],
-                "planRecommendations": [
-                    {
-                        "planRecommendationKey": plan_proposition_instance.id,
-                        "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                        "value": plan_proposition_instance.text_value,
-                        "planThemes": [
-                            "http://uri.suomi.fi/codelist/rytj/kaavoitusteema/code/01",
-                        ],
-                        "recommendationNumber": plan_proposition_instance.ordering,
-                        # TODO: plan recommendation documents to be added.
-                        "periodOfValidity": None,
-                    },
-                ],
-                "letterIdentifier": plan_regulation_group_instance.short_name,
-                "colorNumber": "#FFFFFF",
-            },
-        ],
-        "planRegulationGroupRelations": [
-            {
-                "planObjectKey": land_use_area_instance.id,
-                "planRegulationGroupKey": plan_regulation_group_instance.id,
-            },
-            {
-                "planObjectKey": land_use_point_instance.id,
-                "planRegulationGroupKey": point_plan_regulation_group_instance.id,
-            },
-        ],
-    }
-
-
-@pytest.fixture(scope="function")
-def desired_plan_matter_dict(
-    session: Session,
-    desired_plan_dict: dict,
-    plan_instance: models.Plan,
-    participation_plan_presenting_for_public_decision: codes.NameOfPlanCaseDecision,
-    plan_material_presenting_for_public_decision: codes.NameOfPlanCaseDecision,
-    draft_plan_presenting_for_public_decision: codes.NameOfPlanCaseDecision,
-    plan_proposal_sending_out_for_opinions_decision: codes.NameOfPlanCaseDecision,
-    plan_proposal_presenting_for_public_decision: codes.NameOfPlanCaseDecision,
-    participation_plan_presenting_for_public_event: codes.TypeOfProcessingEvent,
-    plan_material_presenting_for_public_event: codes.TypeOfProcessingEvent,
-    plan_proposal_presenting_for_public_event: codes.TypeOfProcessingEvent,
-    plan_proposal_requesting_for_opinions_event: codes.TypeOfProcessingEvent,
-    presentation_to_the_public_interaction: codes.TypeOfInteractionEvent,
-    decisionmaker_type: codes.TypeOfDecisionMaker,
-) -> dict:
-    """
-    Plan matter dict based on https://github.com/sykefi/Ryhti-rajapintakuvaukset/blob/main/OpenApi/Kaavoitus/Palveluväylä/Kaavoitus%20OpenApi.json
-
-    Constructing the plan matter requires certain additional codes to be present in the database and set in the plan instance.
-
-    Let's 1) write explicitly the complex fields, and 2) just check that the simple fields have
-    the same values as the original plan fixture in the database.
-    """
-
-    return {
-        "permanentPlanIdentifier": "MK-123456",
-        "planType": "http://uri.suomi.fi/codelist/rytj/RY_Kaavalaji/code/11",
-        "name": plan_instance.name,
-        "timeOfInitiation": "2024-01-01",
-        "description": plan_instance.description,
-        "producerPlanIdentifier": plan_instance.producers_plan_identifier,
-        "caseIdentifiers": [plan_instance.matter_management_identifier],
-        "recordNumbers": [plan_instance.record_number],
-        "administrativeAreaIdentifiers": ["01"],
-        "digitalOrigin": "http://uri.suomi.fi/codelist/rytj/RY_DigitaalinenAlkupera/code/01",
-        "planMatterPhases": [
-            {
-                "planMatterPhaseKey": "third_phase_test",
-                "lifeCycleStatus": "http://uri.suomi.fi/codelist/rytj/kaavaelinkaari/code/03",
-                "geographicalArea": desired_plan_dict["geographicalArea"],
-                "handlingEvent": {
-                    "handlingEventKey": "whatever",
-                    "handlingEventType": "http://uri.suomi.fi/codelist/rytj/kaavakastap/code/05",
-                    "eventTime": "2024-02-01",
-                },
-                "interactionEvents": [
-                    {
-                        "interactionEventKey": "whatever",
-                        "interactionEventType": "http://uri.suomi.fi/codelist/rytj/RY_KaavanVuorovaikutustapahtumanLaji/code/01",
-                        "eventTime": {"begin": "2024-02-01", "end": "2024-02-01"},
-                    },
-                ],
-                "planDecision": {
-                    "planDecisionKey": "whatever",
-                    "name": "http://uri.suomi.fi/codelist/rytj/kaavpaatnimi/code/04",
-                    "decisionDate": "2024-02-01",
-                    "dateOfDecision": "2024-02-01",
-                    "typeOfDecisionMaker": "http://uri.suomi.fi/codelist/rytj/PaatoksenTekija/code/01",
-                    "plans": [desired_plan_dict],
-                },
-            },
-        ],
-        # TODO: plan documents, source data etc. non-mandatory fields to be added
-    }
-
+from .conftest import assert_dicts_equal
 
 mock_rule = "random_rule"
 mock_matter_rule = "another_random_rule"
@@ -736,49 +351,6 @@ def client_with_plan_data_in_proposal_phase(
     )
     client.plan_dictionaries = client.get_plan_dictionaries()
     return client
-
-
-def assert_lists_equal(list1: list, list2: list, ignore_keys: Optional[List] = []):
-    """
-    Recursively check that lists have the same items in the same order.
-
-    Optionally, certain keys (e.g. random UUIDs set by the database, our script or
-    the remote Ryhti API) can be ignored when comparing dicts in the lists, because
-    they are not provided in the incoming data.
-    """
-    assert len(list1) == len(list2)
-    for item1, item2 in zip(list1, list2):
-        print(f"comparing values {item1} and {item2}")
-        if isinstance(item1, dict):
-            assert_dicts_equal(item1, item2, ignore_keys=ignore_keys)
-        elif isinstance(item1, list):
-            assert_lists_equal(item1, item2, ignore_keys=ignore_keys)
-        else:
-            assert item1 == item2
-
-
-def assert_dicts_equal(
-    dict1: Mapping, dict2: Mapping, ignore_keys: Optional[List] = []
-):
-    """
-    Recursively check that dicts contain the same keys with same values.
-
-    Optionally, certain keys (e.g. random UUIDs set by the database, our script or
-    the remote Ryhti API) can be ignored when comparing, because they are not
-    provided in the incoming data.
-    """
-    for key in dict2.keys():
-        if not ignore_keys or not key in ignore_keys:
-            assert key in dict1
-    for key, value in dict1.items():
-        if not ignore_keys or not key in ignore_keys:
-            print(f"comparing {key} {value} to {dict2[key]}")
-            if isinstance(value, dict):
-                assert_dicts_equal(dict2[key], value, ignore_keys=ignore_keys)
-            elif isinstance(value, list):
-                assert_lists_equal(dict2[key], value, ignore_keys=ignore_keys)
-            else:
-                assert dict2[key] == value
 
 
 def test_get_plan_dictionaries(
